chemical_equipment_visualiser/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
reportlab==4.0.4
python-dotenv==1.0.0
gunicorn==21.2.0